+++ NEWS.md
@@ -1,5 +1,6 @@
 # cpp11 (development version)
 
+* `cpp_source()` can now accept a source file called `cpp11.cpp` (#133)
 * Writable vector classes now properly set their lengths as intended when being copied to a read only class (#128).
 * `cpp11::named_arg` now explicitly protect their values, avoiding protection issues when using large inputs. [tidyverse/readr#1145](https://github.com/tidyverse/readr/issues/1145)
 * `cpp_source()` now throws an error if the code fails to compile, as intended (#125).